set directory of files
set directory of files
