+++ models/partner.js
