+++ package.json
@@ -4,7 +4,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "dev": " nodemon -r dotenv/config --experimental-json-modules src/index.js",
+    "dev": " nodemon -r dotenv/config --experimental-json-modules index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
